Fixed issue modifying other users notes

+++ website/views.py
@@ -7,6 +7,11 @@ from . import db
 views = Blueprint('views', __name__)
 
 
+@views.route('/manual')
+def manual():
+    return render_template('manual.html', user=current_user)
+
+
 @views.route('/', methods=['GET'])
 @login_required
 def my_notes():
@@ -28,16 +33,22 @@ def post_note():
 @login_required
 def delete_note(note_id):
     note = Note.query.filter_by(id=note_id).first()
+    if note.user_id != current_user.id:
+        flash('You are not allowed to modify this note!', category='error')
+        return redirect(url_for('views.my_notes'))
     if note:
         db.session.delete(note)
         db.session.commit()
     return redirect(url_for('views.my_notes'))
 
 
-@views.route('/done/<note_id>', methods=['GET'])
+@views.route('/check-note/<note_id>', methods=['GET'])
 @login_required
 def check_note(note_id):
     note = Note.query.filter_by(id=note_id).first()
+    if note.user_id != current_user.id:
+        flash('You are not allowed to modify this note!', category='error')
+        return redirect(url_for('views.my_notes'))
     if note:
         note.complete = not note.complete
         db.session.commit()
@@ -50,6 +61,9 @@ def check_note(note_id):
 @login_required
 def edit_note(note_id):
     note = Note.query.filter_by(id=note_id).first()
+    if note.user_id != current_user.id:
+        flash('You are not allowed to modify this note!', category='error')
+        return redirect(url_for('views.my_notes'))
     if note:
         new_content = request.form[f'content{note_id}']
         change = False if note.content.strip() == new_content.strip() else True
@@ -58,10 +72,18 @@ def edit_note(note_id):
         if change:
             flash('Note edited successfully', category='success')
     else:
-        flash('That note does not exist any more', category='error')
+        flash('That note does not exist', category='error')
     return redirect(url_for('views.my_notes'))
 
 
-@views.route('/manual')
-def manual():
-    return render_template('manual.html', user=current_user)
+@views.route('/more_info/<note_id>', methods=['GET', 'POST'])
+@login_required
+def more_info(note_id):
+    note = Note.query.filter_by(id=note_id).first()
+    if note.user_id != current_user.id:
+        flash('You are not allowed to modify this note!', category='error')
+        return redirect(url_for('views.my_notes'))
+    if note:
+        pass
+    else:
+        flash('That note does not exist', category='error')
